Refactored io writers
